1. Changed themes selection animation. 2. Global refactoring
