+++ Calculator2020/Support files/Assets.xcassets/AppIcon.appiconset/Contents.json
