+++ Calculator2020/Support files/Assets.xcassets/empty-frame.imageset/empty-frame.png
