+++ Calculator2020/Support files/Info.plist
